+++ src/Tasks/PermissionProviderBuildTask.php
@@ -37,11 +37,9 @@ class PermissionProviderBuildTask extends BuildTask
     {
         $permissions = Permission::get();
         foreach ($permissions as $permission) {
-            if (0 === $permission->Arg && 1 === $permission->Type) {
-                if (isset($this->_permissions[$permission->Code])) {
-                    DB::alteration_message('Deleting double permission with code: ' . $permission->Code, 'deleted');
-                    $permission->delete();
-                }
+            if (0 === $permission->Arg && 1 === $permission->Type && isset($this->_permissions[$permission->Code])) {
+                DB::alteration_message('Deleting double permission with code: ' . $permission->Code, 'deleted');
+                $permission->delete();
             }
         }
     }
